feat: send task to redis with db transaction
